+++ packages/console-sdk-microfrontend/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @mia-platform/console-sdk-microfrontend
 
+## 0.0.16
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @mia-platform/console-types@0.18.13
+
+## 0.0.15
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @mia-platform/console-types@0.18.12
+
+## 0.0.14
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @mia-platform/console-types@0.18.11
+
+## 0.0.13
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @mia-platform/console-types@0.18.10
+
 ## 0.0.12
 
 ### Patch Changes